Newly updated code after laptop lost

+++ src/main/java/com/load/Ai/dto/ServerHealthDTO.java
@@ -10,13 +10,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ServerHealthDTO {
+
     private Long serverId;
     private String serverName;
     private String serverIp;
     private String status;
     private Double cpuUsage;
     private Double memoryUsage;
-    private Integer currentLoad;
+    private Long currentLoad;              // 🔥 change to Long
     private Integer totalRequestsHandled;
 }
-
